# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: MailBrief API
  description: |
    REST API for MailBrief, a multi-user email digest dashboard.
    Users receive AI-generated daily digests of their newsletters via managed inboxes.
  version: 1.0.0
  contact:
    name: MailBrief
    url: https://mailbrief.io
servers:
  - url: https://mailbrief.io
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /webhook/email:
    post:
      operationId: ingestEmail
      summary: Ingest a raw email
      description: |
        Receives a raw RFC 5322 email from the Cloudflare Email Worker.
        The X-Original-To header identifies the target managed inbox.
      tags: [Webhooks]
      security:
        - WebhookSecret: []
      parameters:
        - name: X-Original-To
          in: header
          required: true
          schema:
            type: string
          description: The managed inbox address the email was sent to.
      requestBody:
        required: true
        content:
          message/rfc822:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Email accepted for processing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/admin/users/{id}/token-limit:
    patch:
      operationId: setUserTokenLimit
      summary: Set user token limit
      description: Overrides the monthly token limit for a specific user.
      tags: [Admin]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [MonthlyTokenLimit]
              properties:
                MonthlyTokenLimit:
                  type: integer
                  minimum: 0
      responses:
        "200":
          description: Token limit updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
  /api/admin/settings/token-limit:
    patch:
      operationId: setDefaultTokenLimit
      summary: Set default token limit
      description: Changes the default monthly token limit applied to new users.
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [DefaultMonthlyTokenLimit]
              properties:
                DefaultMonthlyTokenLimit:
                  type: integer
                  minimum: 0
      responses:
        "200":
          description: Default limit updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from Better Auth (passed as Authorization header).
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Shared secret for Cloudflare Email Worker webhook.
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for cron and service-to-service calls.
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok

    # ── Email ───────────────────────────────────────────────────────
...